+++ gdsfactory/components/spirals/spiral_heater.py
@@ -3,13 +3,12 @@ from __future__ import annotations
 import numpy as np
 
 import gdsfactory as gf
-from gdsfactory.component import Component, ComponentReference
+from gdsfactory.component import Component
 from gdsfactory.components.bends.bend_euler import bend_euler
-from gdsfactory.components.bends.bend_s import bend_s, get_min_sbend_size
+from gdsfactory.components.bends.bend_s import get_min_sbend_size
 from gdsfactory.components.waveguides.straight import straight
 from gdsfactory.routing.route_single import route_single
 from gdsfactory.typings import (
-    ComponentFactory,
     ComponentSpec,
     CrossSectionSpec,
     Floats,
@@ -19,18 +18,16 @@ from gdsfactory.typings import (
 
 @gf.cell
 def spiral_racetrack(
-    min_radius: float = 5,
+    min_radius: float | None = None,
     straight_length: float = 20.0,
     spacings: Floats = (2, 2, 3, 3, 2, 2),
-    straight_factory: ComponentFactory = straight,
-    bend_factory: ComponentFactory = bend_euler,
-    bend_s_factory: ComponentFactory = bend_s,
+    straight: ComponentSpec = straight,
+    bend: ComponentSpec = bend_euler,
+    bend_s: ComponentSpec = "bend_s",
     cross_section: CrossSectionSpec = "strip",
     cross_section_s: CrossSectionSpec | None = None,
-    n_bend_points: int = 99,
-    with_inner_ports: bool = False,
     extra_90_deg_bend: bool = False,
-    allow_min_radius_violation: bool = True,
+    allow_min_radius_violation: bool = False,
 ) -> Component:
     """Returns Racetrack-Spiral.
 
@@ -38,70 +35,47 @@ def spiral_racetrack(
         min_radius: smallest radius in um.
         straight_length: length of the straight segments in um.
         spacings: space between the center of neighboring waveguides in um.
-        straight_factory: factory to generate the straight segments.
-        bend_factory: factory to generate the bend segments.
-        bend_s_factory: factory to generate the s-bend segments.
+        straight: factory to generate the straight segments.
+        bend: factory to generate the bend segments.
+        bend_s: factory to generate the s-bend segments.
         cross_section: cross-section of the waveguides.
         cross_section_s: cross-section of the s bend waveguide (optional).
-        n_bend_points: optional bend points.
-        with_inner_ports: if True, will build the spiral, but expose the inner ports where the S-bend would be.
         extra_90_deg_bend: if True, we add an additional straight + 90 degree bent at the output, so the output port is looking down.
         allow_min_radius_violation: if True, will allow the s-bend to have a smaller radius than the minimum radius.
     """
     c = gf.Component()
 
-    if with_inner_ports:
-        bend_s: ComponentReference | Component = gf.get_component(
-            bend_s_factory,
-            size=(straight_length, -min_radius * 2 + 1 * spacings[0]),
-            cross_section=cross_section_s or cross_section,
-            npoints=n_bend_points,
-        )
-        c.info["length"] = 0
+    xs = gf.get_cross_section(cross_section)
+    min_radius = min_radius or xs.radius
+    assert min_radius
 
-        c.add_port(
-            "o3",
-            center=bend_s.ports["o1"].center,
-            orientation=0,
-            cross_section=bend_s.ports["o1"].cross_section,
-        )
-        c.add_port(
-            "o4",
-            center=bend_s.ports["o2"].center,
-            orientation=180,
-            cross_section=bend_s.ports["o2"].cross_section,
-        )
-    else:
-        _bend_s = gf.get_component(
-            bend_s_factory,
-            size=(straight_length, -min_radius * 2 + 1 * spacings[0]),
-            cross_section=cross_section_s or cross_section,
-            npoints=n_bend_points,
-            allow_min_radius_violation=allow_min_radius_violation,
-        )
-        bend_s = c << _bend_s
-        c.info["length"] = _bend_s.info["length"]
+    _bend_s = gf.get_component(
+        bend_s,
+        size=(straight_length, -min_radius * 2 + 1 * spacings[0]),
+        cross_section=cross_section_s or cross_section,
+        allow_min_radius_violation=allow_min_radius_violation,
+    )
+    bend_s_ref = c << _bend_s
+    c.info["length"] = _bend_s.info["length"]
 
     ports: list[Port] = []
-    for port in bend_s.ports:
+    for port in bend_s_ref.ports:
         for i in range(len(spacings)):
             _bend = gf.get_component(
-                bend_factory,
+                bend,
                 angle=180,
                 radius=min_radius + np.sum(spacings[:i]),
-                p=0,
                 cross_section=cross_section,
-                npoints=n_bend_points,
             )
-            bend = c << _bend
-            bend.connect("o1", port)
+            bend_ref = c << _bend
+            bend_ref.connect("o1", port)
 
             _straight = gf.get_component(
-                straight_factory, length=straight_length, cross_section=cross_section
+                straight, length=straight_length, cross_section=cross_section
             )
-            straight = c << _straight
-            straight.connect("o1", bend.ports["o2"])
-            port = straight.ports["o2"]
+            straight_ref = c << _straight
+            straight_ref.connect("o1", bend_ref.ports["o2"])
+            port = straight_ref.ports["o2"]
 
             c.info["length"] += _bend.info["length"] + _straight.info["length"]
         ports.append(port)
@@ -109,16 +83,14 @@ def spiral_racetrack(
     c.add_port("o1", port=ports[0])
 
     if extra_90_deg_bend:
-        bend = c << gf.get_component(
-            bend_factory,
+        bend_ref = c << gf.get_component(
+            bend,
             angle=90,
             radius=min_radius + np.sum(spacings),
-            p=0,
             cross_section=cross_section,
-            npoints=n_bend_points,
         )
-        bend.connect("o1", ports[1])
-        c.add_port("o2", port=bend.ports["o2"])
+        bend_ref.connect("o1", ports[1])
+        c.add_port("o2", port=bend_ref.ports["o2"])
 
     else:
         c.add_port("o2", port=ports[1])
@@ -130,15 +102,13 @@ def spiral_racetrack_fixed_length(
     length: float = 1000,
     in_out_port_spacing: float = 150,
     n_straight_sections: int = 8,
-    min_radius: float = 5,
+    min_radius: float | None = None,
     min_spacing: float = 5.0,
-    straight_factory: ComponentFactory = straight,
-    bend_factory: ComponentFactory = bend_euler,
-    bend_s_factory: ComponentFactory = bend_s,
+    straight: ComponentSpec = straight,
+    bend: ComponentSpec = "bend_circular",
+    bend_s: ComponentSpec = "bend_s",
     cross_section: CrossSectionSpec = "strip",
     cross_section_s: CrossSectionSpec | None = None,
-    n_bend_points: int = 99,
-    with_inner_ports: bool = False,
 ) -> Component:
     """Returns Racetrack-Spiral with a specified total length.
 
@@ -152,17 +122,17 @@ def spiral_racetrack_fixed_length(
         n_straight_sections: total number of straight sections for the racetrack spiral. Has to be even.
         min_radius: smallest radius in um.
         min_spacing: minimum center-center spacing between adjacent waveguides.
-        straight_factory: factory to generate the straight segments.
-        bend_factory: factory to generate the bend segments.
-        bend_s_factory: factory to generate the s-bend segments.
+        straight: factory to generate the straight segments.
+        bend: factory to generate the bend segments.
+        bend_s: factory to generate the s-bend segments.
         cross_section: cross-section of the waveguides.
         cross_section_s: cross-section of the s bend waveguide (optional).
-        n_bend_points: optional bend points.
-        with_inner_ports: if True, will build the spiral, but expose the inner ports where the S-bend would be.
     """
     c = gf.Component()
 
     xs_s_bend = cross_section_s or cross_section
+    xs = gf.get_cross_section(xs_s_bend)
+    min_radius = min_radius or xs.radius
 
     if np.mod(n_straight_sections, 2) != 0:
         raise ValueError("The number of straight sections has to be even!")
@@ -175,8 +145,8 @@ def spiral_racetrack_fixed_length(
         in_out_port_spacing=in_out_port_spacing,
         min_radius=min_radius,
         spacings=spacings,
-        bend_factory=bend_factory,
-        bend_s_factory=bend_s_factory,
+        bend=bend,
+        bend_s=bend_s,
         cross_section_s_bend=xs_s_bend,
         cross_section=cross_section,
     )
@@ -185,13 +155,11 @@ def spiral_racetrack_fixed_length(
         min_radius=min_radius,
         straight_length=straight_length,
         spacings=spacings,
-        straight_factory=straight_factory,
-        bend_factory=bend_factory,
-        bend_s_factory=bend_s_factory,
+        straight=straight,
+        bend=bend,
+        bend_s=bend_s,
         cross_section=cross_section,
         cross_section_s=cross_section_s,
-        n_bend_points=n_bend_points,
-        with_inner_ports=with_inner_ports,
         extra_90_deg_bend=True,
     )
 
@@ -203,8 +171,10 @@ def spiral_racetrack_fixed_length(
         spiral.mirror_x()
 
     # add a bit more to the spiral racetrack to make the in and out ports be aligned in y
-    in_wg = c << straight_factory(
-        spiral.ports["o1"].dx - spiral.dxmin, cross_section=cross_section
+    in_wg = c << gf.get_component(
+        straight,
+        length=spiral.ports["o1"].dx - spiral.dxmin,
+        cross_section=cross_section,
     )
     if np.mod(n_straight_sections // 2, 2) == 1:
         in_wg.mirror_y()
@@ -226,7 +196,7 @@ def spiral_racetrack_fixed_length(
         spiral.ports["o2"],
         o2_temp,
         straight=straight,
-        bend=bend_factory,
+        bend=bend,
         cross_section=xs_s_bend,
         radius=min_radius,
     )
@@ -245,10 +215,10 @@ def spiral_racetrack_fixed_length(
 def _req_straight_len(
     length: float = 1000,
     in_out_port_spacing: float = 100,
-    min_radius: float = 5,
+    min_radius: float | None = None,
     spacings: Floats = (1.0, 1.0),
-    bend_factory: ComponentFactory = bend_euler,
-    bend_s_factory: ComponentFactory = bend_s,
+    bend: ComponentSpec = bend_euler,
+    bend_s: ComponentSpec = "bend_s",
     cross_section: CrossSectionSpec = "strip",
     cross_section_s_bend: CrossSectionSpec = "strip",
 ) -> float:
@@ -257,17 +227,20 @@ def _req_straight_len(
     Args:
         length: total length of the spiral from input to output ports in um.
         in_out_port_spacing: spacing between input and output ports of the spiral in um.
-        min_radius: smallest radius in um.
+        min_radius: smallest radius in um. Defaults to the radius of the cross-section.
         spacings: spacings between adjacent waveguides.
-        bend_factory: factory to generate the bend segments.
-        bend_s_factory: factory to generate the s-bend segments.
+        bend: factory to generate the bend segments.
+        bend_s: factory to generate the s-bend segments.
         cross_section: cross-section of the waveguides.
         cross_section_s_bend: s bend cross section
     """
     from scipy.interpolate import interp1d
 
-    # "Brute force" approach - sweep length and save total length
+    xs = gf.get_cross_section(cross_section)
+    min_radius = min_radius or xs.radius
+    assert min_radius
 
+    # "Brute force" approach - sweep length and save total length
     lens: list[float] = []
 
     # Figure out the min straight for the spiral so that the inner
@@ -290,9 +263,9 @@ def _req_straight_len(
             min_radius=min_radius,
             straight_length=str_len,
             spacings=spacings,
-            straight_factory=straight,
-            bend_factory=bend_factory,
-            bend_s_factory=bend_s_factory,
+            straight=straight,
+            bend=bend,
+            bend_s=bend_s,
             cross_section=cross_section,
             cross_section_s=cross_section_s_bend,
             extra_90_deg_bend=True,
@@ -319,7 +292,7 @@ def _req_straight_len(
             spiral.ports["o2"],
             c.ports["o2"],
             straight=straight,
-            bend=bend_factory,
+            bend=bend,
             cross_section=cross_section_s_bend,
             radius=min_radius,
         )
@@ -337,9 +310,9 @@ def spiral_racetrack_heater_metal(
     straight_length: float = 30,
     spacing: float = 2,
     num: int = 8,
-    straight_factory: ComponentFactory = straight,
-    bend_factory: ComponentFactory = bend_euler,
-    bend_s_factory: ComponentFactory = bend_s,
+    straight: ComponentSpec = straight,
+    bend: ComponentSpec = bend_euler,
+    bend_s: ComponentSpec = "bend_s",
     waveguide_cross_section: CrossSectionSpec = "strip",
     heater_cross_section: CrossSectionSpec = "heater_metal",
     via_stack: ComponentSpec | None = "via_stack_heater_mtop",
@@ -349,13 +322,13 @@ def spiral_racetrack_heater_metal(
     based on https://doi.org/10.1364/OL.400230 .
 
     Args:
-        min_radius: smallest radius.
+        min_radius: smallest radius. Defaults to the radius of the cross-section.
         straight_length: length of the straight segments.
         spacing: space between the center of neighboring waveguides.
-        num: number.
-        straight_factory: factory to generate the straight segments.
-        bend_factory: factory to generate the bend segments.
-        bend_s_factory: factory to generate the s-bend segments.
+        num: number of loops.
+        straight: factory to generate the straight segments.
+        bend: factory to generate the bend segments.
+        bend_s: factory to generate the s-bend segments.
         waveguide_cross_section: cross-section of the waveguides.
         heater_cross_section: cross-section of the heater.
         via_stack: via stack to connect the heater to the metal layer.
@@ -368,9 +341,9 @@ def spiral_racetrack_heater_metal(
         min_radius,
         straight_length,
         (spacing,) * num,
-        straight_factory,
-        bend_factory,
-        bend_s_factory,
+        straight,
+        bend,
+        bend_s,
         waveguide_cross_section,
     )
 
@@ -397,7 +370,8 @@ def spiral_racetrack_heater_metal(
     )
     heater_bot.dmovey(-spacing * num // 2)
 
-    heater_bend = c << gf.components.bend_circular(
+    heater_bend = c << gf.get_component(
+        bend,
         angle=180,
         radius=min_radius + spacing * (num // 2 + 1),
         cross_section=heater_cross_section,
@@ -443,9 +417,9 @@ def spiral_racetrack_heater_doped(
     straight_length: float = 30,
     spacing: float = 2,
     num: int = 8,
-    straight_factory: ComponentFactory = straight,
-    bend_factory: ComponentFactory = bend_euler,
-    bend_s_factory: ComponentFactory = bend_s,
+    straight: ComponentSpec = straight,
+    bend: ComponentSpec = bend_euler,
+    bend_s: ComponentSpec = "bend_s",
     waveguide_cross_section: CrossSectionSpec = "strip",
     heater_cross_section: CrossSectionSpec = "npp",
 ) -> Component:
@@ -454,13 +428,13 @@ def spiral_racetrack_heater_doped(
     based on https://doi.org/10.1364/OL.400230 but with the heater between the loops.
 
     Args:
-        min_radius: smallest radius in um.
+        min_radius: smallest radius in um. Defaults to the radius of the cross-section.
         straight_length: length of the straight segments in um.
         spacing: space between the center of neighboring waveguides in um.
         num: number.
-        straight_factory: factory to generate the straight segments.
-        bend_factory: factory to generate the bend segments.
-        bend_s_factory: factory to generate the s-bend segments.
+        straight: factory to generate the straight segments.
+        bend: factory to generate the bend segments.
+        bend_s: factory to generate the s-bend segments.
         waveguide_cross_section: cross-section of the waveguides.
         heater_cross_section: cross-section of the heater.
     """
@@ -475,9 +449,9 @@ def spiral_racetrack_heater_doped(
         spacings=(spacing,) * (num // 2)
         + (spacing + 1,) * 2
         + (spacing,) * (num // 2 - 2),
-        straight_factory=straight_factory,
-        bend_factory=bend_factory,
-        bend_s_factory=bend_s_factory,
+        straight=straight,
+        bend=bend,
+        bend_s=bend_s,
         cross_section=waveguide_cross_section,
     )
 
@@ -515,9 +489,11 @@ if __name__ == "__main__":
     # test_length_spiral_racetrack()
 
     # c = spiral_racetrack(cross_section="rib")
+    # c = spiral_racetrack(cross_section="nitride", straight_length=54)
     # c = spiral_racetrack()
-    # c = spiral_racetrack()
-    c = spiral_racetrack_fixed_length()
-    # c = spiral_racetrack_heater_metal()
+    # c = spiral_racetrack_fixed_length()
+    c = spiral_racetrack_heater_metal(
+        waveguide_cross_section="nitride", straight_length=54
+    )
     # c = spiral_racetrack_heater_doped()
     c.show()