+++ tests/routing/test_routing_route_bundle_all_angle.py
@@ -17,14 +17,14 @@ def test_route_bundle_all_angle(
     mmi2.dmove((100, 10))
     mmi2.drotate(30)
 
-    _ = gf.routing.route_bundle_all_angle(
+    routes = gf.routing.route_bundle_all_angle(
         c,
         mmi1.ports.filter(orientation=0),
         [mmi2.ports["o2"], mmi2.ports["o1"]],
     )
-    # if check:
-    #     lengths = {i: route.length for i, route in enumerate(routes)}
-    #     data_regression.check(lengths)
+    if check:
+        lengths = {i: int(route.length) for i, route in enumerate(routes)}
+        data_regression.check(lengths)
 
 
 if __name__ == "__main__":
@@ -42,4 +42,6 @@ if __name__ == "__main__":
         mmi1.ports.filter(orientation=0),
         [mmi2.ports["o2"], mmi2.ports["o1"]],
     )
+    lengths = {i: int(route.length) for i, route in enumerate(routes)}
+    print(lengths)
     c.show()